+++ src/Models/Collection.php
@@ -2,6 +2,7 @@
 
 namespace LaraZeus\Bolt\Models;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -46,7 +47,7 @@ class Collection extends Model
         return null;
     }
 
-    protected static function newFactory(): CollectionFactory
+    protected static function newFactory(): Factory
     {
         return CollectionFactory::new();
     }