+++ src/Models/Section.php
@@ -2,6 +2,7 @@
 
 namespace LaraZeus\Bolt\Models;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -59,7 +60,7 @@ class Section extends Model
         });
     }
 
-    protected static function newFactory(): SectionFactory
+    protected static function newFactory(): Factory
     {
         return SectionFactory::new();
     }